src/requirements.txt
opencv-python
scamp
python-rtmidi
scamp_extensions
google-generativeai